Change pickup distance with strength
Toggle strength now also toggles the pickup distance from 3m (standard) to 12m when toggling strength.

The red hand was annoying me now that there are a lot more pick-up-able objects (especially pallets) in FS22.  This means you have to be close to them to notice, but can still see what is available.